Renamed OpenBuffer to MemReqs. Changed std::uint32_t front-end integers to std::uint64_t. Added PixelFormat::RA_32. Fixed namestyling in Tools namespace. Added a Texas::detail::PNG::FilterType enum for the PNG backend. Got de-indexing of PNG files to a working state.

+++ include/Texas/Dimensions.hpp
@@ -6,9 +6,9 @@ namespace Texas
 {
     struct Dimensions
     {
-        std::uint32_t width;
-        std::uint32_t height;
-        std::uint32_t depth;
+        std::uint64_t width;
+        std::uint64_t height;
+        std::uint64_t depth;
     };
 }
